refactor/#17: 코드리뷰 적용완료
